+++ code/application/game_logic/gun_controller.cpp
@@ -17,7 +17,7 @@ Gun_controller::on_update(const float dt)
 
 
 void
-Gun_controller::on_event(const uint32_t id, const void *data, const size_t size_of_data)
+Gun_controller::on_event(const uint32_t id, const void *data, const uint32_t size_of_data)
 {
   switch(id)
   {
