+++ code/application/game_logic/npc_actor_controller.hpp
@@ -12,7 +12,7 @@ public:
 
   uint32_t              get_rtti() const override { return Component_type::npc_actor_controller; }
   void                  on_update(const float dt) override;
-  void                  on_event(const uint32_t id, const void *data, const size_t size_of_data) override;
+  void                  on_event(const uint32_t id, const void *data, const uint32_t size_of_data) override;
 
 
 }; // class
